+++ axum-debug/tests/fail/attrs.rs
@@ -1,0 +1,6 @@
+use axum_debug::debug_handler;
+
+#[debug_handler(foo)]
+async fn handler() {}
+
+fn main() {}