+++ axum-debug/tests/pass/associated_fn_without_self.rs
@@ -1,0 +1,10 @@
+use axum_debug::debug_handler;
+
+struct A;
+
+impl A {
+    #[debug_handler]
+    async fn handler() {}
+}
+
+fn main() {}