+++ axum-debug/tests/pass/ready.rs
@@ -1,0 +1,9 @@
+use axum_debug::debug_handler;
+use std::future::{Ready, ready};
+
+#[debug_handler]
+fn handler() -> Ready<()> {
+    ready(())
+}
+
+fn main() {}